Enable pedantic lints
This lint warns if a function returns a `Result` but doesn't document
the error conditions.

It only runs for **public** functions, so we can reduce the implications
by using more internal functions.

For the sake of this PR, I left visibility alone, and tried to figure
out which error conditions could occur.

Same as `missing_errors_doc`, but warns if a public function panics
without documenting the panic conditions.

This lint warns if you use `Default::default()` rather than using the
name of the type. This is mostly for readability, but it **does** make
a difference and is easy to keep clean.

```diff
  let client_config = kube::config::Config::from_custom_kubeconfig(
      kube_config,
-     &Default::default(),
+     &KubeConfigOptions::default(),
  )
```

This lint warns if a cast from a larger type to a smaller type which may
truncate the value.

In our case, this was casting `Claims::exp` (a timestamp in number of
seconds since the epoch) from `u64` to `usize`, which could be
truncated. I updated the `exp` field to `u64` to avoid truncation.

I checked [the JWT spec] to check whether the spec mandates a 32-bit
integer for `exp`, but it just defines `NumericData` as:

> A JSON numeric value representing the number of seconds from
> 1970-01-01T00:00:00Z UTC until the specified UTC date/time,
> ignoring leap seconds.

Since JSON numbers are not restricted to 32 bits, I think this is fine.

[the JWT spec]: https://www.rfc-editor.org/rfc/rfc7519#section-4.1.4

Warns if a doc comment has `_`, `::` or CamelCase words in it without
backticks.

This commit adds backticks in these cases, which improves the
readability of the docs.

This lint warns if a public function is missing a `#[must_use]`
attribute. This lint only fires for:

- public functions
- with no mutable arguments
- and don't mutate any statics

This forces the caller to use the return value of the function, rather
than discarding it.

This might be overly strict, but we can start here and see how many
false positives we get.

This lint warns if you could have used a more idiomatic `if let` rather
than a `match`.

The example in this commit is pretty clear.

This (original) version of the code warns:

```rust
let auth_header = match auth_header {
    Some(header) => header,
    None => return Ok(None),
};
```

This is the more idiomatic version, which does not warn:

```rust
let Some(auth_header) = auth_header else {
    return Ok(None);
};
```

This lint warns if you use `.iter()` in the context of a `for` loop,
since it's not necessary.

```diff
- for (key, val) in self.incoming_identity.iter() {
+ for (key, val) in &self.incoming_identity {
```

This lint warns if you call `.to_string()` on something like
`&&String`. The lint's justification is that this code bypasses the
optimized specialized version of `ToString` for `&String` or `&str`.

I don't think this is super-important, but I left it in because:

1. I was trying to see if I could get away with retaining the
  `pedantic` lint category.
2. If we're reliably running the lints, either in our editors or via
   `cargo clippy --fix`, this should be a pretty easy case to keep
   clean.

This commit also adjusted the idiom for receiving a path-like, as part
of other nearby changes caused by pedantic lint failures.

The TL;DR is that you can taked `impl AsRef<Path>`, which will allow
`Path`, but also `&str`. You can then use `path.as_ref()` to get an
`&Path`, and you can use `path.as_ref().to_path_buf()` to get a
`PathBuf` that you can share with other "threads" (i.e. `Send + 'static`
closures).

Some functions return `Result`, even if there are no error conditions.

In some cases, this was done for compatibility with a macro that was
expecting a `Result`-returning function. In those cases, I silenced the
lint and documented the reason an infallible function returns a
`Result`.

In other cases, there was no obvious reason for the `Result`, because it
was a concrete function with no error conditions. In those cases, I
refactored the function to not return a `Result`.